LangChain/SemanticSearch/requirements.txt
python-dotenv
pypdf
langchain-community
langchain_huggingface
langchain-mongodb 
pymongo